requirements.txt
wheel
Django==4.2
MPLn23d
MPImfp
django-bootstrap5
opencv-python
pillow
pandas
matplotlib
seaborn
scikit-learn
djangorestframework
markdown
django-filter
djangorestframework-simplejwt
djoser
ldap3
diff-match-patch
django_cleanup
pybeads
lmfit
GitPython
xgboost
lightgbm
optuna
umap-learn
skl2onnx
onnxmltools
onnxruntime
shap
django-environ
celery
django-redis
django-celery-results
pymupdf
pymupdf4llm
pypdf
Levenshtein
bibtexparser
nbconvert
python-docx
python-pptx
GPyOpt
crossrefapi

# for server operation
mysqlclient
uwsgi
